app/requirements.txt
streamlit == 1.36.0
google-generativeai
google-cloud-bigquery
db_dtypes